bot/requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
exceptiongroup==1.2.0
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
sniffio==1.3.0
typing_extensions==4.8.0
yarl==1.9.3
